research-presentation/requirements_backend.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
torch==2.0.1
ultralytics==8.0.196
requests==2.31.0





